% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{c_to_co2}
\alias{c_to_co2}
\title{Convert mass of carbon to mass of CO2}
\usage{
c_to_co2(x)
}
\arguments{
\item{x}{Mass of carbon atoms.}
}
\value{
Mass of CO2
}
\description{
Converts a mass of carbon atoms in CO2 molecules into a mass of CO2
}
\examples{
c_to_co2(100)

}
