% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{co2_to_c}
\alias{co2_to_c}
\title{Convert mass of CO2 to the equivalent mass of carbon atoms}
\usage{
co2_to_c(x)
}
\arguments{
\item{x}{Mass of CO2.}
}
\value{
Mass of carbon atoms
}
\description{
Converts a mass of CO2 molecules into the equivalent mass of carbon atoms.
}
\examples{
co2_to_c(100)

}
