% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_kaya_explorer.R
\name{kaya_app}
\alias{kaya_app}
\title{Get a shinyApp object for the Kaya explorer app}
\usage{
kaya_app(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{logical variable indicating whether to launch
a web browser and open the URL for the shiny app when starting the
local server.}

\item{...}{other arguments to pass to [`shinyApp`][shiny::shinyApp].}
}
\value{
A shinyApp object that can be run to launch the app.
}
\description{
Generate a shinyApp object for the Kaya explorer app.
}
