% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kayaExplorer.R
\docType{package}
\name{kayaExplorer-package}
\alias{kayaExplorer}
\alias{kayaExplorer-package}
\title{kayaExplorer package}
\description{
A Shiny-based explorer for Kaya identity. Loads data from the kayadata package and
provides an interactive shiny app for exploring the data: choose a country or region and
examine trends in the Kaya variables (population, gross domestic product, primary energy
consumption, and carbon dioxide emissions), compare historical trends to the implied trends
necessary to hit decarbonization policy targets (e.g., 70% reduction of emissions by 2050).
Also allows the user to explore the fuel mix that different countries and regions use to
supply their energy needs.
}
\section{License}{


 The \pkg{kayaExplorer} package is open source licensed under the
 MIT License.
}

\section{Bug reports}{

\itemize{
 \item kayaExplorer issue tracker (\url{https://github.com/jonathan-g/kaya-explorer/issues})
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathan-g/kaya-explorer/}
  \item \url{https://jonathan-g.github.io/kaya-explorer/}
  \item Report bugs at \url{https://github.com/jonathan-g/kaya-explorer/issues/}
}

}
\author{
\strong{Maintainer}: Jonathan Gilligan \email{jonathan.gilligan@vanderbilt.edu}

}
