% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_kaya_explorer.R
\name{launch_kaya_explorer}
\alias{launch_kaya_explorer}
\title{Launch the kayaExplorer app}
\usage{
launch_kaya_explorer(rstudio = getOption("kayaExplorer.rstudio"), ...)
}
\arguments{
\item{rstudio}{Only relevant for RStudio users. The default (\code{FALSE}) is
to launch the app in the user's default web browser rather than RStudio's
pop-up Viewer. Users can change the default to \code{TRUE} by setting the
global option \code{options(kayaExplorer.rstudio = TRUE)}.}

\item{...}{Optional arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
Nothing is returned
}
\description{
Launch the kayaExplorer app in the default web browser. RStudio
users also have the option of launching the app in RStudio's pop-up Viewer.
}
\examples{
\dontrun{
launch_kaya_explorer()
}

}
