% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{set_debug}
\alias{set_debug}
\title{Set debugging flag}
\usage{
set_debug(dbg = NULL)
}
\arguments{
\item{dbg}{Logical debugging flag (`TRUE` or `FALSE`).}
}
\value{
The previous value of the debugging flag.
}
\description{
Sets debugging flag to output debugging info to the console.
}
\examples{
set_debug(TRUE)

}
