% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_kaya_tool.R
\name{kaya_app}
\alias{kaya_app}
\title{Get a shinyApp object for the Kaya tool app}
\usage{
kaya_app(launch.browser = TRUE, ...)
}
\value{
A shinyApp object that can be run to launch the app.
}
\description{
Generate a shinyApp object for the Kaya tool app.
}
