% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya_tool_package.R
\docType{package}
\name{kayatool-package}
\alias{kayatool-package}
\title{ShinyStan interface and shinystan R package}
\description{
kayatool is a Shiny application that allows students to explore the Kaya
identity of many nations around the world.
}
\section{License}{


 The \pkg{kayatool} package is open source licensed under the
 MIT License.
}

\section{Bug reports}{

\itemize{
 \item kayatool issue tracker (\url{https://github.com/gilligan-ees-3310/kayatool/issues})
}
}

